agent-study/pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=0.4.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[[tool.uv.index]]
url ="https://mirrors.aliyun.com/pypi/simple/"
default = true
